fix type

+++ src/main/java/com/example/ploop_backend/domain/map/entity/TrashBin.java
@@ -18,8 +18,8 @@ public class TrashBin {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long latitude;
-    private Long longitude;
+    private double latitude;
+    private double longitude;
     private String imageUrl; // 사진 URL 경로
 
     @ManyToOne(fetch = FetchType.LAZY)
